Public Documentation

Documentation for PrismaConvert.jl's public interface.

Contents

Public Documentation

Contents

Index

Public Interface

Index

PrismaConvert

Main.PrismaConvert.close

Main.PrismaConvert.maketif

Main.PrismaConvert.open

Main.PrismaConvert.plot_band

Main.PrismaConvert.plot_cols

Main.PrismaConvert.plot_linewise

Main.PrismaConvert.plot_rows

Public Interface

PrismaConvert — Module

Main module for PrismaConvert.jl – a Julia package for Prisma's HDF5 products.

Exports

. maketif . open . close . plotband . plotlinewise . plotcols . plotrows

Main.PrismaConvert.open — Function

open(
    f,
    mode
)

Opens a HDF5 file, returns a dictionary containing the opened file

using PrismaConvert
open("./file.h5")

File is opened in read only mode by default, can be opened in other modes by providng mode:

using PrismaConvert
open("./file.h5","w")

Main.PrismaConvert.close — Function

close(
    f
)

Closes all files in the provided files dictionary, returns number of files closed

using PrismaConvert
close(my_dict)

Main.PrismaConvert.maketif — Function

maketif( infile,##NB: infile dev esser già aperto, a diff di prconvert outfile::String; allowederrors = nothing, #array containing ids of errors to cut from product tif source="HCO", PAN=true,#boolean: true-> creates panchromatic tif VNIR=true,#boolean: true-> creates vnir tif SWIR=true, #boolean: true-> creates swir tif FULL=true,#boolean: true-> creates full tif joinpriority="VNIR", #choose wether to overwrite SWIR or VNIR bands in FULL tif overwrite=false, #choose wether to overwrite files selbandsvnir=nothing, #vnir bands to get selbandsswir=nothing, #swir bands to get indexes=nothing, cust_indexes=nothing )

Gets bands from an already opened HDF5 file and saves them in a .tif raster file

using PrismaConvert
maketif(my_dict["file 0"], "out/result")

Main.PrismaConvert.plot_band — Function

plot_band(
    band::Array{T,2}
)

Squares the band, then returns heatmap of the downsized map, columns polot and rows plot

Main.PrismaConvert.plot_linewise — Function

plot_linewise(
    band::Array{T,2}
)

Plots 2d input as a 1d array

plot_linewise(
        cube::Array{T,3}
        )

Returns linewise plot of each band

Main.PrismaConvert.plot_rows — Function

plot_rows(
    band::Array{T,2}
)

Plots plot rows

Main.PrismaConvert.plot_cols — Function

plot_cols
    band::Array{T,2}
)

Plots columns